+++ packages/aesthetic-adapter-typestyle/CHANGELOG.md
@@ -1,7 +1,11 @@
-# 1.0.1
+# 1.1.0
+#### 🚀 New
+* Can now pass dynamic style objects to `transform` (will be akin to inline).
+
+# 1.0.1 - 12/03/17
 #### 🛠 Internal
 * Tested against TypeStyle 1.6.
 
-# 1.0.0
+# 1.0.0 - 12/02/17
 #### 🎉 Release
 * Initial release!